Moved provenance into its own aspect.
