+++ magda-csv-connector/aspect-templates/dcat-dataset-strings.js
@@ -77,8 +77,6 @@ let contactPoint = [
 
 let conformsTo = fuzzy.findClosestFieldThreshold(dataset, 0.5, "data standard");
 
-let creation = extractCreation();
-
 let accessLevel = fuzzy.findClosestFieldThreshold(
     dataset,
     0.5,
@@ -104,7 +102,6 @@ return {
     themes,
     keywords,
     contactPoint,
-    creation,
     accessLevel,
     importance,
     status,
@@ -120,75 +117,6 @@ function extractKeywordList(items) {
         }
     }
     return undefined;
-}
-
-function extractCreation() {
-    let isInternallyProduced = fuzzy.findClosestFieldThreshold(
-        dataset,
-        0.5,
-        "created internally"
-    );
-    if (isInternallyProduced) {
-        isInternallyProduced =
-            fuzzy.similarity(isInternallyProduced, "yes") > 0.5;
-    }
-
-    let mechanism = fuzzy.findClosestFieldThreshold(
-        dataset,
-        0.5,
-        "primary source",
-        "additional information"
-    );
-
-    let sourceSystem = fuzzy.findClosestFieldThreshold(
-        dataset,
-        0.5,
-        "source system",
-        "primary source"
-    );
-
-    let likelihoodOfRelease = fuzzy.findClosestFieldThreshold(
-        dataset,
-        0.5,
-        "likelihood of release"
-    );
-
-    let isOpenData = fuzzy.findClosestFieldThreshold(dataset, 0.5, "open data");
-    if (isOpenData) {
-        isOpenData = fuzzy.similarity(isOpenData, "yes") > 0.5;
-    }
-
-    let affiliatedOrganisation = fuzzy.findClosestFieldThreshold(
-        dataset,
-        0.5,
-        "affiliated organisation"
-    );
-
-    if (affiliatedOrganisation) {
-        affiliatedOrganisation = [affiliatedOrganisation];
-    }
-
-    let creation = undefined;
-
-    if (
-        isInternallyProduced !== undefined ||
-        mechanism !== undefined ||
-        sourceSystem !== undefined ||
-        likelihoodOfRelease !== undefined ||
-        isOpenData !== undefined ||
-        affiliatedOrganisation
-    ) {
-        creation = {
-            isInternallyProduced,
-            mechanism,
-            sourceSystem,
-            likelihoodOfRelease,
-            isOpenData,
-            affiliatedOrganisation
-        };
-    }
-
-    return creation;
 }
 
 function extractInformationSecurity() {
