+++ magda-csv-connector/src/test/csv3.json
@@ -49,12 +49,6 @@
                 "themes": ["666", "#GAMERS", "DATA"],
                 "keywords": ["Operations", "Item", "Yes"],
                 "contactPoint": "Fred Nerk\n\nNuclear Plant",
-                "creation": {
-                    "mechanism": "Yes",
-                    "sourceSystem": "CRM",
-                    "likelihoodOfRelease": "Low",
-                    "affiliatedOrganisation": ["AI"]
-                },
                 "accessLevel": "Non-Public",
                 "importance": "Essential",
                 "status": "Current",
@@ -79,6 +73,12 @@
             },
             "dataset-publisher": {
                 "publisher": "org-connector-ea0436e1fdca43c22e55809b9a4f3475"
+            },
+            "provenance": {
+                "mechanism": "Yes",
+                "sourceSystem": "CRM",
+                "likelihoodOfRelease": "Low",
+                "affiliatedOrganisation": ["AI"]
             }
         },
         "sourceTag": "stag"
@@ -134,10 +134,6 @@
                     "Data generated and provided to us by Data61"
                 ],
                 "contactPoint": "Engineering",
-                "creation": {
-                    "mechanism": "Data generated and provided to us by Data61",
-                    "likelihoodOfRelease": "Restricted"
-                },
                 "accessLevel": "Restricted Public",
                 "importance": "Useful",
                 "informationSecurity": {
@@ -155,6 +151,10 @@
             },
             "dataset-publisher": {
                 "publisher": "org-connector-5d554bc5f3d2cd182cdd0952b1fb87ca"
+            },
+            "provenance": {
+                "mechanism": "Data generated and provided to us by Data61",
+                "likelihoodOfRelease": "Restricted"
             }
         },
         "sourceTag": "stag"
@@ -208,12 +208,6 @@
                 "themes": ["For Official Use Only"],
                 "keywords": ["Compliance", "Item", "For Official Use Only"],
                 "contactPoint": "META STD\n\nGaming",
-                "creation": {
-                    "mechanism": "No",
-                    "sourceSystem": "Recycle bin",
-                    "likelihoodOfRelease": "Low",
-                    "affiliatedOrganisation": ["Machine Learning"]
-                },
                 "accessLevel": "T3",
                 "status": "Retired",
                 "informationSecurity": {
@@ -234,6 +228,12 @@
             },
             "dataset-publisher": {
                 "publisher": "org-connector-967ef2eb34634ba418db94dab610ba6f"
+            },
+            "provenance": {
+                "mechanism": "No",
+                "sourceSystem": "Recycle bin",
+                "likelihoodOfRelease": "Low",
+                "affiliatedOrganisation": ["Machine Learning"]
             }
         },
         "sourceTag": "stag"
