+++ magda-registry-aspects/provenance.schema.json
@@ -1,0 +1,34 @@
+{
+    "$schema": "http://json-schema.org/schema#",
+    "title": "Provenance for datasets or distributions",
+    "type": "object",
+    "properties": {
+        "isInternallyProduced": {
+            "title": "Was it internally produced by the metadata?",
+            "type": "boolean"
+        },
+        "mechanism": {
+            "title": "How was it produced?",
+            "type": "string"
+        },
+        "sourceSystem": {
+            "title": "Source system that was used to produce this metadata.",
+            "type": "string"
+        },
+        "likelihoodOfRelease": {
+            "title": "How likely is it to be released as open data?",
+            "type": "string"
+        },
+        "isOpenData": {
+            "title": "Is it already released as open data?",
+            "type": "boolean"
+        },
+        "affiliatedOrganisation": {
+            "title": "Was this dataset produced collaborating with other organisations?",
+            "type": "array",
+            "items": {
+                "type": "string"
+            }
+        }
+    }
+}
